+++ src/pages/Coinpage/coinpage.styles.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+import { Base } from '../../app.styles'
+
+export const MiddleContainer = styled(Base)`
+width: 70vw;`
